--- a/src/main/java/org/shanerx/tradeshop/enumys/DebugLevels.java
+++ b/src/main/java/org/shanerx/tradeshop/enumys/DebugLevels.java
@@ -41,7 +41,9 @@ public enum DebugLevels {
     NAME_COMPARE(7, Level.WARNING), // 64
     SHULKERS_SUCK(8, Level.WARNING), // 128
     ENCHANT_CHECKS(9, Level.WARNING), // 256
-    OUTPUT(10, Level.WARNING) //512
+    OUTPUT(10, Level.WARNING), //512
+    STATIC_CACHING(11, Level.WARNING) //1024
+
     ;
 
     //position is what value to check for this level in the binary string -1.