--- a/src/main/java/org/shanerx/tradeshop/framework/events/HopperShopAccessEvent.java
+++ b/src/main/java/org/shanerx/tradeshop/framework/events/HopperShopAccessEvent.java
@@ -64,4 +64,13 @@ public class HopperShopAccessEvent extends InventoryMoveItemEvent {
 	public HopperDirection getItemDirection() {
 		return dir;
 	}
+
+	@Override
+	public HandlerList getHandlers() {
+		return handlers;
+	}
+
+	public static HandlerList getHandlerList() {
+		return handlers;
+	}
 }