--- a/src/main/java/org/shanerx/tradeshop/listeners/ShopProtectionListener.java
+++ b/src/main/java/org/shanerx/tradeshop/listeners/ShopProtectionListener.java
@@ -82,7 +82,7 @@ public class ShopProtectionListener extends Utils implements Listener {
             return;
         }
 
-        Boolean skip = plugin.getListManager().canSkipHopper(event.getInitiator().getLocation());
+        Boolean skip = plugin.getDataStorage().getDataCache().canSkipHopper(event.getInitiator().getLocation());
 
         if (skip != null) {
             event.setCancelled(skip);
@@ -99,17 +99,17 @@ public class ShopProtectionListener extends Utils implements Listener {
         try {
             invBlock = (fromHopper ? event.getDestination() : event.getSource()).getLocation().getBlock();
         } catch (NullPointerException ignored) {
-            plugin.getListManager().addSkippableHopper(event.getInitiator().getLocation(), false);
+            plugin.getDataStorage().getDataCache().addSkippableHopper(event.getInitiator().getLocation(), false);
             return;
         }
 
         if (!plugin.getListManager().isInventory(invBlock)) {
-            plugin.getListManager().addSkippableHopper(event.getInitiator().getLocation(), false);
+            plugin.getDataStorage().getDataCache().addSkippableHopper(event.getInitiator().getLocation(), false);
             return;
         }
 
         if (!ShopChest.isShopChest(invBlock)) {
-            plugin.getListManager().addSkippableHopper(event.getInitiator().getLocation(), false);
+            plugin.getDataStorage().getDataCache().addSkippableHopper(event.getInitiator().getLocation(), false);
             return;
         }
 
@@ -118,7 +118,7 @@ public class ShopProtectionListener extends Utils implements Listener {
         boolean isForbidden = !Setting.findSetting(shop.getShopType().name() + "SHOP_HOPPER_" + (fromHopper ? "IMPORT" : "EXPORT")).getBoolean();
         if (isForbidden) {
             event.setCancelled(true);
-            plugin.getListManager().addSkippableHopper(event.getInitiator().getLocation(), true);
+            plugin.getDataStorage().getDataCache().addSkippableHopper(event.getInitiator().getLocation(), true);
             return;
         }
 
@@ -136,7 +136,7 @@ public class ShopProtectionListener extends Utils implements Listener {
         Bukkit.getPluginManager().callEvent(hopperEvent);
         debugger.log("ShopProtectionListener: (TSAF) HopperEvent recovered! ", DebugLevels.PROTECTION);
         event.setCancelled(hopperEvent.isForbidden());
-        plugin.getListManager().addSkippableHopper(event.getInitiator().getLocation(), hopperEvent.isForbidden());
+        plugin.getDataStorage().getDataCache().addSkippableHopper(event.getInitiator().getLocation(), hopperEvent.isForbidden());
         debugger.log("ShopProtectionListener: (TSAF) HopperEvent isForbidden: " + hopperEvent.isForbidden(), DebugLevels.PROTECTION);
     }