--- a/src/main/java/org/shanerx/tradeshop/objects/ShopLocation.java
+++ b/src/main/java/org/shanerx/tradeshop/objects/ShopLocation.java
@@ -68,7 +68,7 @@ public class ShopLocation implements Serializable {
 				world = Bukkit.getWorld(locA[1].replace("-", "_"));
 			if (world == null) {
 				throw new IllegalWorldException("Cannot find world " + locA[1], new WorldlessLocation(x, y, z));
-				// Not to maintainer: do NOT remove this aritificial error, it is supposed to be catched elsewhere
+				// Note to maintainer: do NOT remove this artificial error, it is meant to be caught elsewhere
 				// (Temporary fix for metadata world renaming bug until metadata is removed entirely)
 			}
 
@@ -114,5 +114,9 @@ public class ShopLocation implements Serializable {
 
 	public Location getLocation() {
 		return new Location(world, x, y, z);
+	}
+
+	public String toString() {
+		return serialize();
 	}
 }