--- a/src/main/java/org/shanerx/tradeshop/utils/data/CacheType.java
+++ b/src/main/java/org/shanerx/tradeshop/utils/data/CacheType.java
@@ -27,6 +27,7 @@ package org.shanerx.tradeshop.utils.data;
 
 enum CacheType {
     SHOP,
+    SHOP_LOCATION,
     LINKAGE,
     PLAYER
 }