… src/main/java/org/shanerx/tradeshop/utils/data/Json/LinkageConfiguration.java
… src/main/java/org/shanerx/tradeshop/utils/data/Json/ShopConfiguration.java