@SparklingComet 2022-03-31
Replying to https://github.com/Tradeshop/TradeShop/pull/124#issuecomment-1084092248

1. I am not familiar with the exact meaning of those graphs, what do the red lines mean?
2. When does the sharp increase in the test of the cached trades happen? (Check whether it Is when you start trading?
3. Do we know what those spikes in the hopper tests are caused by? Is that the hopper cache being filled and emptied at regular intervals? We might want to confirm by roughly timing the period there.

If the answer to all of those questions is "yes", we might be looking at something good.