@KillerOfPie 2022-03-31
Realized I didn't have a comparison without the plugin running...

![image](https://user-images.githubusercontent.com/7243280/161151665-b2ffa7af-a2f2-4639-bb8a-28d943a489b2.png)
Pre-Teleport Avg mspT: forgot to check
Post-Teleport Avg mspT: ~60ms
Post-Teleport Peak mspT: ~85ms
Return Avg mspT: ~3ms