@KillerOfPie 2022-04-05
Going to leave this out of 2.5.1 since I want to get that out sooner and maybe add in 2.5.2 or 2.6.0. Right now this does need to be changed to use a combination of the cached items and the old checking method which I want to remove. The only Alternative option I can think of now would be to keep all shop locations and linkage data in memory and have them temporarily cached like now. This would make all checks for finding shops a lot quicker and should be less memory intensive then the hoppers caching just with a constant cost. 

I think I would store in a List<Location shopSign> for Shops and linkage data would be a Map<Location storage block, Location linkedShop>. These would either be in;
- DataCache(since most of the methods to retrieve form them are already there and would need minor changes)  or
- ListManager(this would require making almost duplicate methods to what's is in DataCache)

The ShopCache and PlayerCache would staysince those would still be useful.

@SparklingComet input?