@SparklingComet 2022-04-05
I'd say let's go with DataCache.

This might actually be a viable alternative to the current hopper cache, or at least a unification so that we don't have several caching systems being flung around.

We'd probably still do well to only cache shop data for currently loaded chunks. Since chunks are being loaded/unloaded by the hundreds every second on an active server, we might want to have a mechanism that discards unloaded chunk data every X mins. This might still be an issue though, as many chunks are being loaded every second, so that we would still have  a significant amount of filesystem interactions for potentially unused shops...