@SparklingComet 2022-04-05
This is actually bitter because had we been using SQLite storage all along, this would be significantly less complicated.