@KillerOfPie 2022-04-06
## Previous Timing for the caching branch:
Pre-Teleport Avg mspT: ~3.3ms
Post-Teleport Avg mspT: ~78ms
Post-Teleport Peak mspT: ~150ms
Return Avg mspT: ~2.8ms

## Timing with new changes:
Pre-Teleport Avg mspT: ~3.3ms
Post-Teleport Avg mspT: ~78ms
Post-Teleport Peak mspT: ~90ms
Return Avg mspT: ~2.8ms

Just wanted to point out that this change dropped the peak TPS on the hopper test by 40% and I'm not seeing any noticable changes in memory