@SparklingComet 2022-04-10
> 1. Is being version Independent worth losing our own back support

Why will this be version independent? Are we not using newer Bukkit API features that are not supported by older versions beside persistent data containers etc.?

> 2. Should we make the cross-over in 2.6 and just hold this pr for then since this is where I started to notice the added processing?

I'd still rather introduce caching in v2.5.2 so that reasonable testing was done. If I understood your proposal correctly, the crossover could ideally be implemented as an extension of the currently implemented caching system, at another time (v.2.6.0).

I'm still not convinced about losing out on our "back support". v2.3.0 is way too recent for that...

> 3. Should we post a question on discord for 1?

I don't know if people will understand the implications of this, but we can certainly try. Maybe we could make a new channel in order for people to discuss this feature.